Exposed the checkLocale() function in the API, added more checks for loc
It's critical that the LC_NUMERIC locale be set correctly when parsing
or serializing.

We check locale on each call to serialize() or parse*() unless
"autoFixLocale" is set to false, in which case we only check it once.

autoFixLocale defaults to true.

Note that checkLocale() is not using locks in the optimistic case, for
performance, we don't guard its calls to reading the locale with a lock
since in the optimistic case of the locale being ok, it is actually
thread-safe since nothing is modified.

In the pessimistic case -- where we must set the locale, we do use a
static mutex for setting it.

Note that setlocale is inherently thread-unsafe and even our checks are
no guarantee client code isn't also mucking with this function in some
other thread.  It is what it is.